fix: move periodos endpoint to AtividadeController and use Promise.allSettled for resilient fetch
