+++ front/src/services/turmaService.ts
@@ -1,6 +1,6 @@
 import api from './api';
 
 export const listarPeriodosLetivos = async (): Promise<string[]> => {
-  const response = await api.get<string[]>('/Turma/periodos');
+  const response = await api.get<string[]>('/Atividade/periodos');
   return response.data;
 };
